Split tree_props.rs (747 lines) into subdirectory modules
Break up the monolithic tree_props.rs into focused submodules:
- seasonal_tint.rs (368 lines): seasonal color tinting logic, helpers, and system
- intersection_lamps.rs (231 lines): intersection detection and lamp spawning
- prop_lod.rs (130 lines): LOD visibility system for props
- mod.rs (54 lines): re-exports all pub items and defines TreePropsPlugin

All public API items are re-exported from mod.rs so external code
(e.g. plugin_registration.rs) continues to work unchanged.

Part of #1389
